requirements.txt
Django==4.1.6
crispy-bootstrap5==0.7
django-crispy-forms==2.0
psycopg2-binary==2.9.5
ipython==8.11.0
django-extensions==3.2.1
django_debug_toolbar==3.8.1
Pillow==9.5.0
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.1